Fix several signedness warnings in gio/glib-compile-schemas.c
gio/glib-compile-schemas.c: In function ‘key_state_set_range’:
gio/glib-compile-schemas.c:376:17: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’
  376 |   for (i = 0; i < G_N_ELEMENTS (table); i++)
      |                 ^
gio/glib-compile-schemas.c: In function ‘key_state_serialise’:
gio/glib-compile-schemas.c:714:29: error: comparison of integer expressions of different signedness: ‘gint’ {aka ‘int’} and ‘long unsigned int’
  714 |               for (i = 0; i < size / sizeof (guint32); i++)
      |                             ^

+++ gio/glib-compile-schemas.c
@@ -363,7 +363,7 @@ key_state_set_range (KeyState     *state,
     { 'd',                 "-inf",                  "inf" },
   };
   gboolean type_ok = FALSE;
-  gint i;
+  gsize i;
 
   if (state->minimum)
     {
@@ -705,7 +705,7 @@ key_state_serialise (KeyState *state)
               guint32 *words;
               gpointer data;
               gsize size;
-              gint i;
+              gsize i;
 
               data = state->strinfo->str;
               size = state->strinfo->len;
